fix: realizar Mejoras a Recordatorio; (A) Modificar los argumentos de la Entidad Recordatorio en .sql (B) Ajuste de relaciones (C) Refinamiento de mapeo en RecordatorioController (/api/v1/recordatorio) (D) Ajustar de Long a Integer (E) Progreso en implementación de clave primaria compuesta (RecordatorioPK), (F) Cambio temporal en Entidad Usuario con relacion de`OneToMany` en List<Recordatorio>, (G) Eliminar duplicado de Entidad Recordatorio

+++ src/main/java/com/petwellness/service/RecordatorioService.java
@@ -5,5 +5,5 @@ import com.petwellness.dto.RecordatorioDTO;
 public interface RecordatorioService {
     RecordatorioDTO crearRecordatorio(RecordatorioDTO recordatorioDTO);
 
-    void eliminarRecordatorio(Long id);
+    void eliminarRecordatorio(Integer id);
 }
